+++ CHANGELOG.md
@@ -24,6 +24,7 @@ The format is based on [Keep a Changelog][keepachangelog] and this project adher
 
 ### Fixed
 
+- Fix add chat for birthday worker personal
 - Fix execute calculate rule for schedule payments
 
 ## v0.0.1 - 2020-04-15